refactor: add curse inside curse module
