+++ apps/api/src/modules/curse/curse.service.ts
