+++ apps/api/src/route.ts
@@ -1,4 +1,4 @@
-import { curseRoutes } from './modules/curse.controller'
+import { curseRoutes } from './modules/curse/curse.controller'
 import type { FastifyTypedInstance } from './types/fastify'
 
 export async function routes(app: FastifyTypedInstance) {
